backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pymongo[srv]
motor
scikit-learn
lightgbm
numpy
pandas
joblib
google-auth
